login_qt/requirements.txt
mysql_connector_repackaged==0.3.1
PySide6==6.7.1
PySide6==6.7.1
PySide6_Addons==6.7.1
PySide6_Essentials==6.7.1
python_bcrypt==0.3.2
